Upgrade polkadart & its dependencies in signer & tfchain, fix upgrade issue, produce new palletes, edit tfchain connect method

+++ packages/tfchain_client/lib/src/contracts.dart
@@ -22,7 +22,7 @@ class QueryContracts {
     return res;
   }
 
-  Future<List<int>> getActiveContracts({required int nodeId}) async {
+  Future<List<BigInt>> getActiveContracts({required int nodeId}) async {
     final res =
         await client.api.query.smartContractModule.activeNodeContracts(nodeId);
     return res;
